Improve CIN description and colors to be more accurate
Hey, I tried to email this to hello@ppg.report but it bounced. I studied meteorology and noticed the last sentence of the CIN description (and chosen colors) is wrong because when CIN is closer to 0, it is more dangerous to fly. (Zero inhibition means there is nothing stopping air from convecting). 

I've also updated the CIN ranges used for weak, moderate, and strong to reflect what I've seen in scientific papers. 

It's confusing because of the negative values used to describe negative energy, but it's important for pilots to know that CIN values closer to zero are more dangerous and CIN values farther from zero are more stable. This all depends on CAPE as well.

I am NOT a programmer at all and made a GitHub account just for this.